+++ operators-examples/openshift-qiskit-operator/image/Dockerfile
