+++ operators-examples/openshift-qiskit-operator/image/jupyter_notebook_config.py
