+++ operators-examples/openshift-qiskit-operator/operator/deploy/crds/singhp11.io_v1_qiskitplayground_cr.yaml
@@ -1,0 +1,7 @@
+apiVersion: singhp11.io/v1
+kind: QiskitPlayground
+metadata:
+  name: example-qiskitplayground
+spec:
+  # Add fields here
+  size: 3
