+++ operators-examples/openshift-qiskit-operator/operator/deploy/operator.yaml
@@ -1,0 +1,33 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: openshift-qiskit-operator
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      name: openshift-qiskit-operator
+  template:
+    metadata:
+      labels:
+        name: openshift-qiskit-operator
+    spec:
+      serviceAccountName: openshift-qiskit-operator
+      containers:
+        - name: openshift-qiskit-operator
+          # Replace this with the built image name
+          image: singhp11/openshift-qiskit-operator:demo
+          command:
+            - openshift-qiskit-operator
+          imagePullPolicy: Always
+          env:
+            - name: WATCH_NAMESPACE
+              valueFrom:
+                fieldRef:
+                  fieldPath: metadata.namespace
+            - name: POD_NAME
+              valueFrom:
+                fieldRef:
+                  fieldPath: metadata.name
+            - name: OPERATOR_NAME
+              value: "openshift-qiskit-operator"
