+++ operators-examples/openshift-qiskit-operator/operator/deploy/secret.cfg
