+++ operators-examples/openshift-qiskit-operator/operator/deploy/sectret.yaml
