+++ operators-examples/openshift-qiskit-operator/operator/deploy/service_account.yaml
@@ -1,0 +1,4 @@
+apiVersion: v1
+kind: ServiceAccount
+metadata:
+  name: openshift-qiskit-operator
