+++ operators-examples/openshift-qiskit-operator/operator/pkg/apis/addtoscheme_singhp11_v1.go
@@ -1,0 +1,10 @@
+package apis
+
+import (
+	v1 "github.com/husky-parul/openshift-qiskit-operator/pkg/apis/singhp11/v1"
+)
+
+func init() {
+	// Register the types with the Scheme so the components can map objects to GroupVersionKinds and back
+	AddToSchemes = append(AddToSchemes, v1.SchemeBuilder.AddToScheme)
+}
