+++ operators-examples/openshift-qiskit-operator/operator/pkg/apis/singhp11/group.go
@@ -1,0 +1,6 @@
+// Package singhp11 contains singhp11 API versions.
+//
+// This file ensures Go source parsers acknowledge the singhp11 package
+// and any child packages. It can be removed if any other Go source files are
+// added to this package.
+package singhp11
