+++ operators-examples/openshift-qiskit-operator/operator/pkg/apis/singhp11/v1/doc.go
@@ -1,0 +1,4 @@
+// Package v1 contains API Schema definitions for the singhp11 v1 API group
+// +k8s:deepcopy-gen=package,register
+// +groupName=singhp11.io
+package v1
