+++ operators-examples/openshift-qiskit-operator/operator/pkg/apis/singhp11/v1/qiskitplayground_types.go
