+++ operators-examples/openshift-qiskit-operator/operator/version/version.go
