--- operators-examples/qiskit-operator/README.md
@@ -1,8 +1,0 @@
-# Artifacts Detail
-
-    .
-        ├── image                   # Files required for building Jupyter Notebook image
-        ├── operator                # Operator artifacts and documentation 
-        └── README.md
-
-To learn how to deploy the operator on OpenShift 4 follow operator [README](https://github.com/husky-parul/glowing-quantum/tree/qiskit-dev-op/operators-examples/qiskit-dev-operator/operator).
