--- operators-examples/qiskit-operator/etc/jdob-operator-source.yaml
@@ -1,11 +1,0 @@
-apiVersion: operators.coreos.com/v1
-kind: OperatorSource
-metadata:
-  name: jdob-operators
-  namespace: openshift-marketplace
-spec:
-  type: appregistry
-  endpoint: https://quay.io/cnr
-  registryNamespace: jdob
-  displayName: "Jason Dobies Testing Operators"
-  publisher: "Jason Dobies"
