--- operators-examples/qiskit-operator/operator/deploy/crds/dobtech.io_qiskitplaygrounds_crd.yaml
@@ -1,42 +1,0 @@
-apiVersion: apiextensions.k8s.io/v1beta1
-kind: CustomResourceDefinition
-metadata:
-  name: qiskitplaygrounds.dobtech.io
-spec:
-  group: dobtech.io
-  names:
-    kind: QiskitPlayground
-    listKind: QiskitPlaygroundList
-    plural: qiskitplaygrounds
-    singular: qiskitplayground
-  scope: Namespaced
-  subresources:
-    status: {}
-  validation:
-    openAPIV3Schema:
-      description: QiskitPlayground is the Schema for the qiskitplaygrounds API
-      properties:
-        apiVersion:
-          description: 'APIVersion defines the versioned schema of this representation
-            of an object. Servers should convert recognized schemas to the latest
-            internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources'
-          type: string
-        kind:
-          description: 'Kind is a string value representing the REST resource this
-            object represents. Servers may infer this from the endpoint the client
-            submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds'
-          type: string
-        metadata:
-          type: object
-        spec:
-          description: QiskitPlaygroundSpec defines the desired state of QiskitPlayground
-          type: object
-        status:
-          description: QiskitPlaygroundStatus defines the observed state of QiskitPlayground
-          type: object
-      type: object
-  version: v1
-  versions:
-  - name: v1
-    served: true
-    storage: true
