--- operators-examples/qiskit-operator/operator/deploy/crds/dobtech.io_v1_qiskitplayground_cr.yaml
@@ -1,4 +1,0 @@
-apiVersion: dobtech.io/v1
-kind: QiskitPlayground
-metadata:
-  name: example-playground
