--- operators-examples/qiskit-operator/operator/deploy/olm-catalog/qiskit-operator/manifests/qiskit-playground.package.yaml
@@ -1,5 +1,0 @@
-packageName: qiskit-playground-operator
-channels:
-- name: stable
-  currentCSV: qiskit-playground-operator.v1.0.0
-defaultChannel: stable
