--- operators-examples/qiskit-operator/operator/deploy/operator.yaml
@@ -1,32 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: qiskit-operator
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      name: qiskit-operator
-  template:
-    metadata:
-      labels:
-        name: qiskit-operator
-    spec:
-      serviceAccountName: qiskit-operator
-      containers:
-        - name: qiskit-operator
-          image: singhp11/qiskit-operator
-          command:
-          - qiskit-operator
-          imagePullPolicy: Always
-          env:
-            - name: WATCH_NAMESPACE
-              valueFrom:
-                fieldRef:
-                  fieldPath: metadata.namespace
-            - name: POD_NAME
-              valueFrom:
-                fieldRef:
-                  fieldPath: metadata.name
-            - name: OPERATOR_NAME
-              value: "qiskit-operator"
