--- operators-examples/qiskit-operator/operator/deploy/role_binding.yaml
@@ -1,11 +1,0 @@
-kind: RoleBinding
-apiVersion: rbac.authorization.k8s.io/v1
-metadata:
-  name: qiskit-operator
-subjects:
-- kind: ServiceAccount
-  name: qiskit-operator
-roleRef:
-  kind: Role
-  name: qiskit-operator
-  apiGroup: rbac.authorization.k8s.io
