--- operators-examples/qiskit-operator/operator/deploy/service_account.yaml
@@ -1,4 +1,0 @@
-apiVersion: v1
-kind: ServiceAccount
-metadata:
-  name: qiskit-operator
