--- operators-examples/qiskit-operator/operator/pkg/apis/addtoscheme_dobtech_v1.go
@@ -1,10 +1,0 @@
-package apis
-
-import (
-	v1 "github.com/jdob/qiskit-operator/pkg/apis/dobtech/v1"
-)
-
-func init() {
-	// Register the types with the Scheme so the components can map objects to GroupVersionKinds and back
-	AddToSchemes = append(AddToSchemes, v1.SchemeBuilder.AddToScheme)
-}
