--- operators-examples/qiskit-operator/operator/pkg/apis/dobtech/group.go
@@ -1,6 +1,0 @@
-// Package dobtech contains dobtech API versions.
-//
-// This file ensures Go source parsers acknowledge the dobtech package
-// and any child packages. It can be removed if any other Go source files are
-// added to this package.
-package dobtech
