--- operators-examples/qiskit-operator/operator/pkg/apis/dobtech/v1/doc.go
@@ -1,4 +1,0 @@
-// Package v1 contains API Schema definitions for the dobtech v1 API group
-// +k8s:deepcopy-gen=package,register
-// +groupName=dobtech.io
-package v1
